# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: ReviewFlow Pro API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /feedback:
    get:
      operationId: listFeedback
      tags: [feedback]
      summary: List all negative feedback
      parameters:
        - name: clientId
          in: query
          schema:
            type: integer
        - name: priority
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of feedback
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaginatedFeedback"
  /feedback/{id}:
    get:
      operationId: getFeedback
      tags: [feedback]
      summary: Get feedback by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Feedback found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Feedback"

    patch:
      operationId: updateFeedback
      tags: [feedback]
      summary: Update feedback status/priority
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateFeedbackBody"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Feedback"

  # ─── Billing / Subscriptions ───
...
  /notifications:
    get:
      operationId: listNotifications
      tags: [notifications]
      summary: List notifications for current user
      parameters:
        - name: read
          in: query
          schema:
            type: boolean
        - name: page
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of notifications
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaginatedNotifications"
  /notifications/{id}/read:
    patch:
      operationId: markNotificationRead
      tags: [notifications]
      summary: Mark a notification as read
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Marked as read
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Notification"
...
  /public/review/{qrCode}:
    get:
      operationId: getPublicReviewPage
      tags: [public]
      summary: Get review page data for a QR code
      parameters:
        - name: qrCode
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Review page data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PublicReviewPageData"
        "404":
          description: QR code not found or expired
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /public/review/{qrCode}/submit:
    post:
      operationId: submitPublicReview
      tags: [public]
      summary: Submit a review or feedback from the public flow
      parameters:
        - name: qrCode
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SubmitReviewBody"
      responses:
        "200":
          description: Submission result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SubmitReviewResponse"
  /public/review/{qrCode}/scan:
    post:
      operationId: recordQrScan
      tags: [public]
      summary: Record a QR code scan event
      parameters:
        - name: qrCode
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RecordScanBody"
      responses:
        "200":
          description: Scan recorded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required: [error]
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required: [success]

    # ─── Auth ───
...
    Feedback:
      type: object
      properties:
        id:
          type: integer
        campaignId:
          type: integer
        clientId:
          type: integer
        clientName:
          type: string
        rating:
          type: integer
        feedbackText:
          type: ["string", "null"]
        complaintCategory:
          type: ["string", "null"]
        customerName:
          type: ["string", "null"]
        customerPhone:
          type: ["string", "null"]
        customerEmail:
          type: ["string", "null"]
        priority:
          type: string
          enum: [low, medium, high, urgent]
        status:
          type: string
          enum: [new, in_progress, resolved, closed]
        deviceType:
          type: ["string", "null"]
        createdAt:
          type: string
      required: [id, campaignId, clientId, clientName, rating, priority, status, createdAt]
    UpdateFeedbackBody:
      type: object
      properties:
        priority:
          type: string
          enum: [low, medium, high, urgent]
        status:
          type: string
          enum: [new, in_progress, resolved, closed]
    PaginatedFeedback:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Feedback"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
      required: [data, total, page, limit]

    # ─── Subscription ───
...
    Notification:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        title:
          type: string
        message:
          type: string
        type:
          type: string
          enum: [review, feedback, billing, system, alert]
        read:
          type: boolean
        entityId:
          type: ["integer", "null"]
        entityType:
          type: ["string", "null"]
        createdAt:
          type: string
      required: [id, userId, title, message, type, read, createdAt]
    PaginatedNotifications:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Notification"
        total:
          type: integer
        unreadCount:
          type: integer
        page:
          type: integer
        limit:
          type: integer
      required: [data, total, unreadCount, page, limit]

    # ─── Analytics ───
...
    PublicReviewPageData:
      type: object
      properties:
        campaignId:
          type: integer
        clientId:
          type: integer
        businessName:
          type: string
        logoUrl:
          type: ["string", "null"]
        recommendedKeywords:
          type: array
          items:
            type: string
        businessCategory:
          type: string
        isActive:
          type: boolean
      required: [campaignId, clientId, businessName, recommendedKeywords, businessCategory, isActive]
    SubmitReviewBody:
      type: object
      properties:
        rating:
          type: integer
        reviewText:
          type: ["string", "null"]
        keywords:
          type: array
          items:
            type: string
        feedbackText:
          type: ["string", "null"]
        complaintCategory:
          type: ["string", "null"]
        customerName:
          type: ["string", "null"]
        customerPhone:
          type: ["string", "null"]
        customerEmail:
          type: ["string", "null"]
        deviceType:
          type: ["string", "null"]
      required: [rating]
    SubmitReviewResponse:
      type: object
      properties:
        type:
          type: string
          enum: [positive, negative]
        googleReviewUrl:
          type: ["string", "null"]
        reviewText:
          type: ["string", "null"]
        message:
          type: string
      required: [type, message]
    RecordScanBody:
      type: object
      properties:
        deviceType:
          type: ["string", "null"]
        userAgent:
          type: ["string", "null"]
        location:
          type: ["string", "null"]